Game class
The game class handles all of the UI components such as layout and interaction.

It's first method, ShowScreen creates an instance of a JFrame which will serve as the main screen for the game and the JLabel which displays the number that will increase/decrease through the player clicking on buttons.

The inclusion of the JFont will allow allow for all UI elements to be more visible as the font's default size is too small.