Better UI layout and button hold
Added separate JPanels for the number label and buttons to allow for better positioning and size of the elements as there is also dedicated fonts for them (with the number font being bigger to make it be the primary focus on screen).

Also implemented a way for the buttons to be held down with the mouse which can increase/decrease the number consecutively through the usage of MouseAdapters and Timers.

The MouseAdapters includes helpful overridable methods which are invoked depending on the mouses action with the button such as mousePressed and mouseReleased whilst the timer allows for multiple mouse events to be fired.